src/orchestrator/pyproject.toml
[project]
name = "opal-orchestrator"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.1",
  "azure-identity>=1.17",
  "azure-servicebus>=7.12",
  "azure-storage-blob>=12.20",
  "httpx>=0.27",
  "tenacity>=8.3",
  "pydantic>=2.7",
  "pydantic-settings>=2.2",
  "structlog>=24.1",
  "opal-shared @ file:///app/shared"
]

[tool.setuptools.packages.find]
where = ["."]
